Repo updates

+++ inc/MillerRabin.hxx
@@ -66,7 +66,7 @@
 		}
 
 		for (int a : {2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37}) {
-			if (n == a)
+			if (n == (u64)a)
 				return true;
 			if (check_composite(n, a, d, r))
 				return false;
